+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 - Allow setting table `OPTIONS` using `config` ([#171](https://github.com/fishtown-analytics/dbt-spark/pull/171))
+- Add support for column comment ([#170](https://github.com/fishtown-analytics/dbt-spark/pull/170))
+
 
 ### Fixes
 
@@ -17,6 +19,8 @@
 - [@franloza](https://github.com/franloza) ([#160](https://github.com/fishtown-analytics/dbt-spark/pull/160))
 - [@Fokko](https://github.com/Fokko) ([#165](https://github.com/fishtown-analytics/dbt-spark/pull/165))
 - [@JCZuurmond](https://github.com/JCZuurmond) ([#171](https://github.com/fishtown-analytics/dbt-spark/pull/171))
+- [@cristianoperez](https://github.com/cristianoperez) ([#170](https://github.com/fishtown-analytics/dbt-spark/pull/170))
+
 
 ## dbt-spark 0.19.1 (Release TBD)